% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LithoGas-package.R
\docType{package}
\name{LithoGas-package}
\alias{LithoGas}
\alias{LithoGas-package}
\title{LithoGas: Monte Carlo Modelling for Geologic Hydrogen and Helium}
\description{
Provides Monte Carlo modelling of hydrogen and helium generation rates via radiolysis and serpentinization models, extended to economic estimations incorporating source area volume plots and deep time production estimates. Implements radiolysis methods of Warr et al. (2023) \doi{10.3389/feart.2023.1150086} and serpentinization methods of Ardakani et al. (2026) \doi{XXXXXXXX}.
}
\author{
\strong{Maintainer}: Daniel Coutts \email{daniel.coutts@nrcan-rncan.gc.ca}

Authors:
\itemize{
  \item Oliver Warr
  \item Omid Ardakani
  \item Barbara Sherwood Lollar
}

}
\keyword{internal}
