% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteSerpFeTotal.R
\name{monteSerpFeTotal}
\alias{monteSerpFeTotal}
\title{monteSerpFeTotal}
\usage{
monteSerpFeTotal(DF, numGen)
}
\arguments{
\item{DF}{one row of a structured sample dataframe}

\item{numGen}{Number of Monte Carlo simulations}
}
\value{
A data frame with \code{numGen} rows containing computed
  serpentinization variables. Input columns from \code{DF} are removed;
  only newly computed columns are returned for binding by
  \code{\link{monteProd}}. Key output columns include:
  \describe{
    \item{\code{SerpMolH2Rate}}{Hydrogen production rate
      (mol H\eqn{_2} / m\eqn{^3} rock / year).}
    \item{\code{SerpMassH2Rate}}{Hydrogen production rate
      (kg H\eqn{_2} / m\eqn{^3} rock / year).}
    \item{\code{SerpMolH2}}{Cumulative hydrogen production
      (mol H\eqn{_2} / m\eqn{^3} rock) over sample age.}
    \item{\code{SerpMassH2}}{Cumulative hydrogen mass
      (kg H\eqn{_2} / m\eqn{^3} rock) over sample age.}
    \item{\code{SerpModel}}{Character label: \code{"Fe Total Serp"} to denote allow for filtering of results to one type of model or another.}
    }
}
\description{
Monte Carlo model of hydrogen production through serpentinization, from total iron data

Computes hydrogen production rates via serpentinization for a set of
Monte Carlo trials, using total iron concentration (Fe\eqn{_2}O\eqn{_3}\eqn{_T}.
The change in Fe\eqn{^{3+}}/Fe\eqn{_T} ratio between initial
(pre-serpentinization) and current states is used to estimate the mass of
magnetite (Fe\eqn{_3}O\eqn{_4}) produced and the associated hydrogen yield.
}
\details{
The monteSerpFeTotal() function takes a single row of a structured dataframe and computes H2 production rate. Parameters are grabbed by column name.
}
\examples{
data("structuredDF")
monteProd(structuredDF[1,],50,serp=TRUE,allowTotalFeSerp=TRUE)

}
\references{
Ardakani, O.A., Sherwood Lollar, B., Coutts, D.S., Warr, O.A., et al.
(2026).
}
\seealso{
\code{\link{monteProd}}, \code{\link{monteSerpFeSpecies}},
  \code{\link{rtrunc}}
}
